requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
async-generator==1.10
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2023.10
Django==5.0.1
django-ckeditor==6.7.0
django-core==1.4.1
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==3.0.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-secret-manager==2.17.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
